Ball Physics + Minor Tweaking
Added Physics Mat for the ball (in the Assets folder) giving the ball bounce.

Slightly unnecessary but added another "wall" (floor) to the court, previously sometimes when I loaded in I fell through the floor, added another layer of collision... maybe helps?

Also in scene slightly raised player (to help with falling)

Tweaked with the ball holding so it actively follows the "holdpoint" gameobject connected with player... So these are very sensitive and moving the holdpoint w/out consideration will mess with the ball holding...