Added simple Start screen
Created start screen with buttons to start game or quit game.
Made a second button for second version of the game maybe will implement later if not will delete,
added a new spotlight since I dunno why when i start it onyl one part is light without the second spotlight,
added collision to the rim.
Will work on a in game menu but since we have a timer I will check how to stop that while paused. Unsure if fork or commit, so just trying fork.